timesheet pdf print button fixed now

+++ build_and_deploy_env.sh
@@ -5,18 +5,20 @@
 
 set -x
 
-###### Set some variables
-HOST_ENDPOINT_UNSECURE_URL="http://localhost:8078"
-WORKSPACE_DIR="/tmp/webimpetus/"
-
-##### Set some variables
-
 if [[ -z "$1" ]]; then
    echo "env is empty, so setting target_env to development (default)"
    target_env="development"
 else
    echo "env is NOT empty, so setting target_env to $1"
    target_env=$1
+fi
+
+if [[ -z "$2" ]]; then
+   echo "action is empty, so setting action to start (default)"
+   env_action="start"
+else
+   echo "action is NOT empty, so setting action to start (default)"
+   env_action=$2
 fi
 
 target_env_short=$target_env
@@ -29,24 +31,51 @@ if [[ "$target_env" == "development" ]]; then
 target_env_short="dev"
 fi
 
+###### Set some variables
+HOST_ENDPOINT_UNSECURE_URL="http://localhost:8078"
+
+##### Set some variables
+WORKSPACE_DIR=$(pwd)
+
+if [[ "$target_env_short" == "test" ]]; then
+WORKSPACE_DIR="/tmp/webimpetus"
 mkdir -p ${WORKSPACE_DIR}
 chmod 777 ${WORKSPACE_DIR}
-
 rm -rf ${WORKSPACE_DIR}*
-cp -r ../webimpetus/* ${WORKSPACE_DIR}
+cp -r ../webimpetus/* ${WORKSPACE_DIR}/
+fi
 
-mv ${WORKSPACE_DIR}${target_env_short}.env ${WORKSPACE_DIR}.env
-
-cd ${WORKSPACE_DIR}
-
-docker-compose -f docker-compose.yml down
-docker-compose -f docker-compose.yml up -d --build
-docker-compose -f docker-compose.yml ps
+if [[ "$target_env_short" == "prod" ]]; then
+WORKSPACE_DIR="/tmp/webimpetus"
+mkdir -p ${WORKSPACE_DIR}
+chmod 777 ${WORKSPACE_DIR}
+rm -rf ${WORKSPACE_DIR}*
+cp -r ../webimpetus/* ${WORKSPACE_DIR}/
+fi
 
 if [[ "$target_env_short" == "dev" ]]; then
+echo "No need to move dev env files"
+else
+mv ${WORKSPACE_DIR}/${target_env_short}.env ${WORKSPACE_DIR}/.env
+fi
+cd ${WORKSPACE_DIR}/
+
+if [[ "$env_action" == "stop" ]]; then
+docker-compose -f "${WORKSPACE_DIR}/docker-compose.yml" down
+fi
+
+if [[ "$env_action" == "start" ]]; then
+docker-compose -f "${WORKSPACE_DIR}/docker-compose.yml" down
+docker-compose -f "${WORKSPACE_DIR}/docker-compose.yml" up -d --build
+docker-compose -f "${WORKSPACE_DIR}/docker-compose.yml" ps
+fi
+
+if [[ "$target_env_short" == "dev" && "$env_action" == "start" ]]; then
 chmod +x reset_containers.sh
 /bin/bash reset_containers.sh $target_env
 fi
+
+if [[ "$env_action" == "start" ]]; then
 
 sleep 2
 
@@ -63,7 +92,7 @@ if [[ "$os_type" == "Linux" ]]; then
 xdg-open $HOST_ENDPOINT_UNSECURE_URL
 fi
 
-
+fi
